% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IO.R
\name{arr2df}
\alias{arr2df}
\title{extract data from array into a dataframe for selected locations and substances,}
\usage{
arr2df(arr, locmod, submod)
}
\arguments{
\item{arr}{the array to be extracted.}

\item{locmod}{the locations in the array to be extracted}

\item{submod}{the substances in the array to be extracted}
}
\value{
A dataframe with model output values for \code{submod} and \code{locmod}.
}
\description{
extract data from array into a dataframe for selected locations and substances,
}
\examples{
library(Waternet)
arr <- sobek2arr(filename = "data/testdata.his")
submod <- c("OXY", "Cl")
locmod <- c("LOX003","LOX009")
df <- arr2df(arr, locmod=locmod, submod=submod)
library(ggplot2)
plot <- ggplot(df, aes(time, value)) +
  geom_line(aes(color = variable), size = 1) +
  facet_grid((variable ~ location), scales = "free")
plot
}
