% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datafunctions.R
\name{cum_values}
\alias{cum_values}
\title{Create cumulative plots for dicharge series,}
\usage{
cum_values(df, debug = F)
}
\arguments{
\item{df}{the input dataframe, containing <time>, <location>, <variable>, <value> values}

\item{debug}{flag for debug output, default = FALSE}
}
\value{
An R dataframe containing cumnulative values, based on value * dt.
}
\description{
Function to compute cumulative values, i,e. cumulative volumes based on flow values (m3/s) and the dt (s) in the dataset.
Cum values are resetted every year.
}
\examples{
library(Waternet)
submod <- c("OXY")
locmod <- c("LOX010")
df <- get_model_data("DATA/testdata.his", locmod, submod)
df.cum <- cum_values(df, debug=T)
library(ggplot2)
plot <- ggplot(df.cum, aes(time, value.cum)) +
  geom_line(aes(color = variable), size = 1) +
  facet_grid((variable ~ location), scales = "free")
plot
}
