% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IO.R
\name{delwaq2arr}
\alias{delwaq2arr}
\title{convert Delwaq <.his> or <.map> file into a R array object}
\usage{
delwaq2arr(filename, timestamp = T, begintime = "1900-01-01 00:00:00")
}
\arguments{
\item{filename}{the <.his> or <.map> file to be converted.}
}
\value{
An R array object of the Delwaq <.his> or <.map> file named  \code{filename}.
}
\description{
This function reads a delwaq binairy result file with extention <.map> or <.his> and converts it to a R array object.
}
\details{
convert Delwaq <.map> or <.his> file into an R array object,
}
\examples{
library(Waternet)
arr <- delwaq2arr(filename = "DATA/testdata.his")
submod <- c("OXY", "Cl")
locmod <- c("LOX003","LOX009")
df <- arr2df(arr, locmod=locmod, submod=submod)
library(ggplot2)
plot <- ggplot(df, aes(time, value)) +
  geom_line(aes(color = variable), size = 1) +
  facet_grid((variable ~ location), scales = "free")
plot
}
