% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IO.R
\name{get_data_locs}
\alias{get_data_locs}
\title{Get locations from data array,}
\usage{
get_data_locs(data)
}
\arguments{
\item{'data}{object (array).}
}
\value{
list of locations in data
}
\description{
Get locations from data array,
}
