% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IO.R
\name{get_data_tims}
\alias{get_data_tims}
\title{Get times from data array,}
\usage{
get_data_tims(arr)
}
\arguments{
\item{arr}{data object (array).}
}
\value{
list of times in data
}
\description{
Get times from data array,
}
