% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IO.R
\name{get_his_vars}
\alias{get_his_vars}
\title{get SOBEK <.map> or <.his> file variable names,}
\usage{
get_his_vars(filename)
}
\arguments{
\item{filename}{the <.his> or <.map> file.}
}
\value{
A vector object of the variable names.
}
\description{
get SOBEK <.map> or <.his> file variable names,
}
\examples{
library(Waternet)
vars <- get_his_vars("DATA/testdata.his")
}
