% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IO.R
\name{get_model_data}
\alias{get_model_data}
\title{Get model results for different modelruns,}
\usage{
get_model_data(filename, locs, vars)
}
\arguments{
\item{filename}{the name of the file containg the binairy output}

\item{locs}{list of locations to be be extracted}

\item{vars}{list of variables to be be extracted}
}
\value{
A dataframe with model output values for \code{submod} and \code{locmod}.
}
\description{
Function to get modelresults for the same location and variables for multiple runs
}
\examples{
library(Waternet)
submod <- c("OXY", "Cl")
locmod <- c("LOX003","LOX009")
df <- get_model_data("data/testdata.his", locmod, submod)
library(ggplot2)
plot <- ggplot(df, aes(time, value)) +
  geom_line(aes(color = variable), size = 1) +
  facet_grid((variable ~ location), scales = "free")
plot
}
