% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/massbalance.R
\name{hrt_swt_meteotype}
\alias{hrt_swt_meteotype}
\title{Compute HRT per Mass balance area per meteotype}
\usage{
hrt_swt_meteotype(df.bal, f.meteotype = NULL)
}
\arguments{
\item{df.bal}{the mass balance input dataframe, containing <stof>, <tag>, <month>, <year>, <location>,<meteotype>, <meteotype_label>,
<richting>, <volume_m3>}

\item{f.subs}{filter for substances, by default set to "all" to process all variables available in the <variable> column}
}
\value{
An R dataframe HRT per Mass balance area
}
\description{
Function to compute the HRT in a mass balance area per meteotype
Meteotypes are periods of more or less equal hydrological conditions, ie dry summers, wet summer, etc.
}
