% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/massbalance.R
\name{hrt_swt_month}
\alias{hrt_swt_month}
\title{Compute HRT per Mass balance area per month}
\usage{
hrt_swt_month(df.bal, f.meteotype = NULL)
}
\arguments{
\item{df.bal}{the mass balance input dataframe, containing <stof>, <tag>, <month>, <year>, <location>
<richting>, <volume_m3>}

\item{f.subs}{filter for substances, by default set to "all" to process all variables available in the <variable> column}
}
\value{
An R dataframe HRT per Mass balance area
}
\description{
Function to compute the HRT in a mass balance area per month
}
