% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/massbalance.R
\name{loads_in_swt_meteotype}
\alias{loads_in_swt_meteotype}
\title{Aggregate loads per Mass balance area, term, substance and meteo type}
\usage{
loads_in_swt_meteotype(df.bal, f.subs = "all", f.meteotype = NULL)
}
\arguments{
\item{df.bal}{the mass balance input dataframe, containing <stof>, <tag>, <month>, <year>, <location>, <meteotype>, <meteotype_label>,
<term>, <richting>, <surface_m2>}

\item{f.subs}{filter for substances, by default set to "all" to process all variables available in the <variable> column}

\item{f.meteotype}{filter for meteotypes, as integer}
}
\value{
An R dataframe containing loads per balance term
}
\description{
Function to compute the load in a mass balance area per meteotype
Meteotypes are periods of more or less equal hydrological conditions, ie dry summers, wet summer, etc.
}
