% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/massbalance.R
\name{loads_in_swt_month}
\alias{loads_in_swt_month}
\title{Aggregate loads per Mass balance area, term, substance and month}
\usage{
loads_in_swt_month(df.bal, f.subs = "all")
}
\arguments{
\item{df.bal}{the mass balance input dataframe, containing <stof>, <tag>, <month>, <year>, <location>,
<term>, <richting>, <surface_m2>}

\item{f.subs}{filter for substances, by default set to "all" to process all variables available in the <variable> column}
}
\value{
An R dataframe containing loads per balance term
}
\description{
Function to compute the load in a mass balance area on a monthly interval
}
