% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/massbalance.R
\name{merge_bal_periods}
\alias{merge_bal_periods}
\title{Merge Balance data with meteotype labels.}
\usage{
merge_bal_periods(df.bal, df.periods)
}
\arguments{
\item{df.bal}{the mass balance input dataframe, containing <stof>, <tag>, <month>, <year>, <location>
<richting>, <volume_m3>}

\item{df.periods}{meteotype dataframe containing <delwaq_time>.}
}
\description{
Specicific for Waternet:
Function to merge the mass balance dates with the meteotype classification
}
