% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/massbalance.R
\name{merge_bal_terms}
\alias{merge_bal_terms}
\title{Merge Balance data with mass balance labels.}
\usage{
merge_bal_terms(df.bal, df.balterms)
}
\arguments{
\item{df.bal}{the mass balance input dataframe, containing <stof>, <tag>, <month>, <year>, <location>
<richting>, <volume_m3>}

\item{df.balterms}{mass balance labels containing <variable>, <term>, <term_basis>, <stof>, <lokaal>, 
<transport>, <richting>, <type>}
}
\description{
Specicific for Waternet:
Function to merge the mass balance data with the mass balance labels
}
