% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IO.R
\name{OpenHISFile}
\alias{OpenHISFile}
\title{opens the map file, reads the header and some constants}
\usage{
OpenHISFile(filename)
}
\arguments{
\item{filename}{the filename of the map file to be read.}
}
\value{
A dataframe with header and some constants of \code{filename}.
}
\description{
opens the map file, reads the header and some constants
}
