% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsers.R
\name{parse_DELWAQ_monitor_file}
\alias{parse_DELWAQ_monitor_file}
\title{read DELWAQ monitor.dat inlcude file into an R dataframe object,
the file contains the monitoring area's as aplied in a Sobek model.}
\usage{
parse_DELWAQ_monitor_file(filename)
}
\arguments{
\item{filename}{monitoring file to be read.}
}
\value{
R dataframe containing the folowing columns: .
}
\description{
read DELWAQ monitor.dat inlcude file into an R dataframe object,
the file contains the monitoring area's as aplied in a Sobek model.
}
