% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsers.R
\name{parse_DELWAQ_pointer_file}
\alias{parse_DELWAQ_pointer_file}
\title{read DELWAQ pointer.dat file into an R dataframe object,
the file the relation between the delwaq segments.}
\usage{
parse_DELWAQ_pointer_file(filename)
}
\arguments{
\item{filename}{DELWAQ pointer file to be read.}
}
\value{
R dataframe containing the folowing columns: 1) from segment number,2) to segment number.
}
\description{
read DELWAQ pointer.dat file into an R dataframe object,
the file the relation between the delwaq segments.
}
