% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IO.R
\name{ReadHISFile}
\alias{ReadHISFile}
\title{opens the map file, reads map output data for a number of substances for one time slice}
\usage{
ReadHISFile(
  MFS,
  Subs = seq(1, MFS$NumSubs),
  HSegms = seq(1, MFS$NumSegm/MFS$NumLay),
  VSegms = seq(1, MFS$NumLay),
  Time = 1
)
}
\arguments{
\item{MFS}{A dataframe with information on the map file, returned from OpenMapFile\(\).}
}
\value{
A matrix with model map output values for \code{MFS}.
}
\description{
opens the map file, reads map output data for a number of substances for one time slice
}
