% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leafletMap.R
\name{RotatedMarker.layer}
\alias{RotatedMarker.layer}
\title{Add a leaflet layer containing rotated arrows:}
\usage{
RotatedMarker.layer(map, data, unit, grouplayer, bins, plot.angle = T)
}
\arguments{
\item{map}{leaflet map as base layer}

\item{data}{dataframe containg the folowing columns: 
1) value [numeric] 
2) angle in degrees (N = 0, clockwise) [numeric] 
3) name of the location [character]
4) lat of the location [numeric] 
5) lon of the location  [numeric]}

\item{unit}{of the value's plotted on the map}

\item{bins}{bins for the colormap used}

\item{plot.angle}{(default TRUE) of plotting direction of the value}

\item{grouplater}{as layer option for showing/hiding the map layer}
}
\value{
leaflet map object.
}
\description{
Add a leaflet layer containing rotated arrows:
}
