% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datafunctions.R
\name{wqmeas_swt_meteotype}
\alias{wqmeas_swt_meteotype}
\title{Compute average concentrations per (mass balance) area per meteotype}
\usage{
wqmeas_swt_meteotype(df.map, df.wq, df.period, swt, f.meteotype = NULL)
}
\arguments{
\item{df.wq}{the df containing WQ measurement data, containing <time>, <location>, <variable>, <value> values}

\item{df.period}{dataframe contating information regarding}

\item{f.meteotype}{filter for meteotypes, as integer}

\item{map.wqloc_swt}{mapping table with the relation between measurment location and (mass balance) area.}

\item{debug}{flag for debug output, default = FALSE}

\item{f.subs}{filter for substances, by default set to "all" to process all variables available in the <variable> column}
}
\value{
An R dataframe containing cumnulative values, based on value * dt.
}
\description{
Function to compute the average concentrations per (mass balance) area per meteotype
Meteotypes are periods of more or less equal hydrological conditions, ie dry summers, wet summer, etc.
}
