% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IO.R
\name{get_data_vars}
\alias{get_data_vars}
\title{Get varabeles from data array,}
\usage{
get_data_vars(data)
}
\arguments{
\item{'data}{object (array).}
}
\value{
list of variables in data
}
\description{
Get varabeles from data array,
}
